requirements.txt
# SHawn-BIO Dependencies
# Python 3.9+ required

# ===================
# Core Data Science
# ===================
pandas>=2.0.0
numpy>=1.24.0

# ===================
# Bioinformatics
# ===================
biopython>=1.81

# ===================
# Machine Learning
# ===================
scikit-learn>=1.3.0

# ===================
# Vector Search (SBI Pipeline)
# ===================
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# ===================
# Document Processing
# ===================
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
pypdf>=3.17.0

# ===================
# Logging
# ===================
loguru>=0.7.0

# ===================
# Async Support
# ===================
# asyncio - built-in in Python 3.9+

# ===================
# Optional: Environment Config
# ===================
python-dotenv>=1.0.0

# ===================
# Optional: Jupyter Notebook Support
# ===================
jupyter>=1.0.0
ipykernel>=6.0.0

# ===================
# Optional: Visualization
# ===================
matplotlib>=3.7.0
seaborn>=0.12.0
